LPS-55026 consistency
